Update task with max value

+++ Task1_Max_value.py
@@ -1,6 +1,6 @@
-val1 = int(input("Enter first value: "))
-val2 = int(input("Enter second value: "))
-val3 = int(input("Enter third value: "))
+val1 = float(input("Enter first value: "))
+val2 = float(input("Enter second value: "))
+val3 = float(input("Enter third value: "))
 
 max_val = val1
